break up large modules
